Implements xpset
