+++ src/g_asm.cpp
@@ -29,50 +29,7 @@ void xsetmode(void) {
 void xshowpage(unsigned page) {
 
 }
-void xline(int16_t x0, int16_t y0, int16_t x1, int16_t y1, int16_t page, int16_t color) {
-  int16_t x, y;
 
-  if (x0 == x1 && y0 == y1)
-    xpset(x0, y0, page, color);
-
-  else if (std::abs(x1 - x0) >= std::abs(y1 - y0))
-  {
-    if (x1 < x0)
-    {
-      x = x1;
-      y = y1;
-      x1 = x0;
-      y1 = y0;
-      x0 = x;
-      y0 = y;
-    }
-    for (x = x0; x <= x1; x++)
-    {
-      y = (int)(y0 + ((x - x0) * (int32_t)(y1 - y0)) / (x1 - x0));
-      xpset(x, y, page, color);
-    }
-  }
-  else
-  {
-    if (y1 < y0)
-    {
-      x = x1;
-      y = y1;
-      x1 = x0;
-      y1 = y0;
-      x0 = x;
-      y0 = y;
-    }
-    for (y = y0; y <= y1; y++)
-    {
-      x = (int)(x0 + ((y - y0) * (int32_t)(x1 - x0)) / (y1 - y0));
-      xpset(x, y, page, color);
-    }
-  }
-}
-void xpset(int16_t X, int16_t Y, uint16_t PageBase, int16_t Color) {
-
-}
 int16_t xpoint(int16_t X, int16_t Y, uint16_t PageBase) {
   return 0;
 }
